Improve the GenerateType code-actions menu experience.
Put generate-type actions under a higher level "Generate Type" menu item if there are a lot of items in the code actions list.
If there aren't a lot of actions, then inline the items.
Simplify the wording of the "Generate type" actions to make the menu items cleaner.
Provide simple APIs to allow users to create top level menu actions with sub actions.