Relax check for generated files.
In the previous commit, the check for generated files was a bit tight -
The generate method codefix won't even show up if there is not a
destination that is not a generated file. This commit relaxes that
requirement a little bit. It offers the codefix, but "prefers" a
non-generated file over a generated file. If generated file is the only
available destination, then it is chosen.