For multiple implementations, show the location and qualified name
This requires some updated handling of our INavigableItem interface.
The Visual Studio implementation previously would only show the
DisplayString if it had children, and would otherwise show file
location and the source snippet. For implementations, the source
snippet is somewhat useless: in the case of methods you just see
the method signature a bunch of times. Thus, we wanted to show source
location and a fully qualified name, something which wasn't expressible.

The new interface now has DisplayString, which can be null if you want
the string to be line from the source. Whether you want to show the file
location is explicitly called out now too.