Speed up smart indenting in the presence of many inactive regions
Avoid repeatedly calling ISyntaxFactsService.IsInInactiveRegion, which calls SyntaxTree.GetToken. GetToken needs to traverse directive trivia to find the token it is
attached to. This results in n ^ 2 traversals of the tree if there are repeated directive before the line to indent. Additionally, when the indenter determines
that it's examining a line inside a disabled region, instead of examining the previous line, it examines the first line preceeding the current line that isn't
inside a disabled region.

Fixes #5394.