Add null check for TextViewModel in GetDataTipText...
The Web Tools team reported seeing occasional crashes in DataTips.  Dump analysis pointed to GetDataTipText being called on a WpfTextView that was already closed.  Even if the caller (VS IDE) were to check this, there is a race where some other thread could close the view before GetDataTipText executes.  We need to handle this and return gracefully.