Enable GenerateConversionTest tests
Related to #5422

Due to a limitation of xUnit, the existence of the test method
TestEquivalenceKey() and its overload test helper
TestEquivalenceKey(string, string) causes the test runner to bail on
running any test in the containing type of GenerateConversionTest.
Changing the helper method's name to TestEquivalenceKeyWorker(...)
should get these tests running again.