Increase reliability of compiler server tests
Rather than introducing wait time for the server to start up, this
change makes the compiler server tests wait for the existence of the
server mutex. Also, instead of waiting a fixed time for the server to
exit, simply wait up to a given time and finish the test if the server
exits earlier (thus speeding up the test in the fast case).