Fix indentation when pressing ENTER in a partially completed from clause of a query expression
In Visual Studio 2013, we would always indent to the start of the 'from' keyword when pressing ENTER inside of a from clause. For example:

```C#
var q = from x
        |
```

and

```C#
var q = from x in
        |
```

Unfortunately, we regressed that behavior in Visual Studio 2015. It now does this:

```C#
var q = from x
			|
```

This change restores the VS 2013 behavior by adding a special indent block operation specifically for ```FromClauseSyntax```. However, it also continues to support nested query formatting, which we improved in VS 2015:

```C#
var q = from x in
		    from y in args
		    select y
		select x;
```