Unbreak End Construct around an broken code scenario
If you hit enter after a Property statement that already had accessors,
but was missing an End Property, we'd get confused of what to do and
not let you do anything at all. That's bad, so now we just do nothing
special and pass it through as an enter. In this case spitting anything
just makes more broken code for the user, so we're best off avoiding
any fanciness.

Fixes #5208.