Fix generate constructor in a derived type with an abstract base.
Generate constructor had a logic bug causing it to consider any single constructor
in an abstract base class to be an acceptable base-class-constructor for it to
delegate to.  This was the case, regardless of what the actual parameter types of
that constructor were.  Now we actually consider the parameter types in this case
and only allow delegation if the parameter types would be assignable.