Fix timeout in AsyncOperationListenerTests
These tests were essentially deadlocking because the current scheduler
here is the StaTaskScheduler.  When the original thread hit the Wait
call the inlining code in the scheduler allowed the Task created in the
method body to run inline on the original thread.  This quickly lead to
a deadlock as both threads ended up waiting for each other in a blocking
fashion.

Fixed this by forcing the Task in the method body to run in the thread
pool.