Don't explicitly focus the editor when our ICompletionSession is dismissed
Years ago, we added a line of code to explicitly focus the editor when our ICompletionSession is dismissed, to handle a specific problem with VB completion. Essentially, through a subtle confluence of focus issues and events firing, VB line commit would run when double-clicking an item in the completion list to commit it. However, this line is causing numerous other problems in Visual Studio. (There's a particularly bad one when clicking VS toolbar item when the completion list is up.) After removing this line, I can no longer reproduce the original problem. So, I'm removing this to address the other problems. If it turns out that the original problem still occurs in certain circumstances, we'll work with the VS editor team to add a proper fix.