Break dependency on prerelease metadata feature
The signature decoder out of the dev/metadata branch will not ship in v1.1
RTM of System.Reflection.Metadata. Breaking changes to the API are coming
and the API will be officially reviewed and only checked in to corefx master
after v1.1 has snapped.

This temporarily forks the relevant code from corefx dev/metadata in to the
Roslyn tree as internal types. It sets things up so that dotnet/corefx#3560
being merged to master will be enough to get Roslyn using -rc builds out of
corefx master instead of -alpha builds out of corefx dev/metadata.