Introduce Project.HasCompleteReferencesAsync
When you get a compilation from a project in a workspace, it's possible
that the compilation doesn't have all the references you expect. If the
project has references to other languages, we have to produce reference
assemblies, a process that might fail. If it fails, your compilation
just doesn't get references you thought it'd get. Now with this API,
you can ask if the compilation is going to have such missing references,
and act accordingly.

The API is internal for now until get some additional testing on it and
verify it's good for our immediate needs.