Back out the configure block change
This change produced XML that was malformed based on what the pull request build plugin was expecting.  This was causing an exception in the pull request builder, cancelling pull request processing.  I modified the plugin to avoid the exception, but it still appears the Roslyn requests are malformed.

I am backing this out for now until we can understand what is really going on.