Unify restore across Windows and Linux
This changes the Windows builds to restore in the same manner as Linux.  Helps ensure we keep the lab consistent between the runs.

The other positive upside of this change is shaves ~10 minutes of individual runs.  Roslyn.sln is a *huge* solution, has a rather large MSBuild include pattern and as a result puts a lot of pressure on restore.  As such the set of restores we do typically takes ~10 minutes.

Long term we'll be working with NuGet to get this time down a bit.  At the moment though we're a bit of an odd solution here with our size and time requirements.