Added Restore.cmd to simplify restore for contributors
Restore.cmd can now be called to simplify the number of
projects you need to restore until we've resolved #6336.

This also reduces the number of restores we do during CI
builds, as we were restoring Roslyn and the samples twice
and the toolset 3 times.

Make note, I removed -nocache. This does *nothing*
for project.json projects, it does not stop NuGet from
using the Global Packages folder
(%USERPROFILE%\.nuget\packages) and it does not
validate this folder's contents with what's in the source.