Disable AppDomain isolation in unit tests
This change disables AppDomain isolation in the unit tests that don't require it.  This is necessary for our goal of running many of the suites on CoreCLR and helps reduce our exposure to #6358.

The AppDomain creation code incorrectly used AppDomain.CurrentDomain.BaseDirectory as the new base directory.  There is no gurantee this points to the directory containing the unit test binaries and this is actually the case in xunit without AppDomains.  Instead it points to the directory containing xunit.oconsole.x86 and uses AppDomain.ResolveAssembly / LoadFrom to resolve unit test binaries.

Switching over to no AppDomains in xunit was causing the creation of RuntimeAssemblyManager across AppDomains to fail.  The error given was a failure to load Roslyn.Test.Utilities.Desktop in the original AppDomain.  This is odd given that the code is executing inside Roslyn.Test.Utilities.Desktop.  The error was fixed by hooking AppDomain.AssemblyResolve in the original AppDomain and essentially reloading the DLL.  I'm not completely sure why this worked but I suspect it has to do with the original AppDomain having an incorrect base directory.