Fix for issue 6425: Right click "Suppress" should be enabled even when the error list selection has some suppressed issues
Scenario: User bulk selects diagnostics in the error list for suppression/baselining, which includes at least one suppressable diagnostic, but still doesn't see any context menu item for suppressing the diagnostics.

Cause: Our current logic to show the Suppress/Remove suppresion commands require that all the selected diagnostics are active/suppressed respectively. The design team has decided that we should loosen this check to "require that at least one of the selected diagnostics is active/suppressed respectively - this decision was based on the feedback from few people dogfooding this feature.

Fix: Trivial fix to not account for the irrelevant diagnostics when deciding whether or not to show the command. The implementation of the command already drops all the irrelevant diagnostics.

Testing: Verified that we get the desired context menu visiblity for mixed error list selection and invoking the commands also work as expected.

Fixes #6425