Open documents with FileShare.ReadWrite | FileShare.Delete
A long time ago, we did this to ensure that when we're asynchronously
reading files we don't lock the file and break other tools trying to
write. We would check the last modified timestamp to know if we
"definitely lost" the race and would try to bail. This behavior was
accidentally lost in 97862d2, and the comment saying why we did that
later lost in cdca065. This restores opening files with ReadWrite |
Delete, and changes what we do if we detect the race. Previously we'd
return the contents of the file, but nothing would then retry the read
later since at the time we had no mechanism. Since then we now treat
IOException as an indication that we need to retry, so we now just
throw IOException when we know something went wrong.