Optimize data passing
Optimize the data passing across AppDomains to not pass ModuleData for
already loaded Assembly values.  ModuleData is relatively expensive to
marshal because it has the full byte image of the Assembly.  No need to
marshal this data for pre-loaded assemblies like mscorlib, System.Core,
etc ...  It is sufficient to verify the ModuleDataId is already present.