Fix issue with VB preprocessor keywords appearing in the completeion list within incomplete preprocessor
To determine whether preprocessor keywords should appear in the completion list, we first check to see if
we're in a processor start context. As part of this check we test for implicit line continuation, which can
be a bit complex. Essentially, we check the trailing trivia of the starting token and each missing token
following it. Then, when we encounter a non-missing token, we check its leading trivia. Unfortunately, the
code didn't expect that the leading trivia of the next non-missing token could overlap with the starting token,
which can happen if the starting token inside of structured trivia.

Consider the following code:

Class C

    #Region |
End Class

In this case, there is a missing string literal token at the marked position to check. Then, we encounter a
non-missing token: "End". However, the leading trivia of "End" includes the incomplete #Region preprocessor
plus the whitespace before it. However, the code never checks to see if the leading trivia overlaps the #Region,
so it happily starts checking for implicit line contination before #Region, causing an incorrect result.

This fix adds that check.