Greatly improve performance of Object Browser first open
Object Browser was forcing compilations to be realized just to compute the assembly identity for each of a compilation's metadata references. With this change, the Object Browser will use the metadata references from the project and call the same utility function that is used to read assembly identity for analyzer assemblies.