Fix MutexStopsServerStarting suite
This suite was flaky because it relied on timeouts and cross process
communication.  The latter in particular was problematic because it
depended on being able to control the lifetime of VBCSCompiler.exe.
This is a process frequently killed in our scripts as it can lock DLLs
and is always safe to kill.  Fixed the flakiness by refactoring the APIs
and eliminating the timeouts