Always "fail fast" in EE unit tests...
We always want to "fail fast" if there is an Exception in the EE unit
tests (we never want to go down the "generate a non-fatal Watson" code
path).  Also, there were several places where we were generating
non-fatal Watsons for cases that were not truly "exceptional" (the
Exception was being handled reasonably and returned as an "error" result).

We should only have a filter to generate a non-fatal Watson in places
where it would also make sense to throw an unhandled Exception.