Remove source code dependencies on our OfficialBuild flag
The official build flag specifies which assembly version is baked into
Roslyn. This was being used in various ways to either include or not
include binding redirects, but we now always want binding redirects.
Also, we were passing the flag along to the compiler to figure out
which NewVersion to bake into the redirects, but setting nothing
at all conveniently does exactly what we want.