Exit on assert failure in bootstrap build
This changes the compilers and server to exit the process when a Debug.Assert
call fails in a bootstrap build.  The current behavior of throwing up a dialog
blocks the Jenkins build until a timeout occurs and leaves no data.  Now a full
stack trace will be printed, the process exits with a non-zero exit code and the
entire run will fail promptly.