.editorconfig: Don't set indent_size for all files
Roslyn has many different kinds of files in it, so it would be a lot of
work to try to assign appropriate indent_size values to all of them, and
would tend to get out of date; consequently, it's not a very good idea
to include indent_size in the [*] section.

One example that was wrong here is .json; .xml and .xsd were also
similarly afflicted. (And .xsd most certainly does not need any more
afflictions beyond those that were "designed" into it.)