Fix Code Model's GetTypeFromFullName to better handle situations where unsupported name formats are passed
`GetTypeFromFullName` first attempts to treat the specified type name as a metadata name. Failing
that, it parses the type name and tries to speculatively bind it. However, it was entirely possible
that a type name could be provided that we don't support, such as an assembly-qualified name. So, we
now check to see if the parsed type name includes an skipped token trivia (e.g. a trailing "Foo, 0.0.0.0")
and does not attempt to bind in that case. Otherwise, speculative binding could return something
non-sensical.