Fix test issues
This fixes as couple of test issues:

- CultureAttributeMismatch: this was a flaky suite whose flakiness was
fixed via the PEVerify changes.  The hardened rules revealed the suite
was setup to consistently fail verification because it didn't fully load
it's dependencies into the Compilation object.  This behavior is
actually expected, we want to get the compiler errors that result from
such a compilation.  Hence we disable verification in those cases.
- Return07_2: This was directly fixed by the PEVerify hardenning.

closes #4300
closes #5866