Enforce that WpfFact tests must run serially
For now, WpfTestCase, when running tests, sets various bits of global
static state. Because of this, it's not safe to let two tests interleave
on the same test thread, and so we should explicitly prevent this.

This change we should revert in the hypothetical future where we can
run multiple WpfFact threads in parallel.