Remove supports sections that are currently causing warnings
These supports sections (as best I can tell) should be here, but since
we aren't including a reference to Microsoft.NETCore.Targets, the
supports targets are unknown and thus just giving warnings. We should
revert this and fix the underlying issue, but after a half hour I
only have more questions than answers. So this isn't great, but it's
better than not doing this.