Ensure that an EnvDTE.CodeParameter.Name returns the correct value for parameters defined in metadata
The issue is that the symbol display system will not return a name for a parameter unless
SymbolDisplayParameterOptions.IncludeName is set. This behavior seems like it might ultimately be
a bug in symbol display as that option is intended to control the display the parameters owned
by the targeted symbol, but probably shouldn't affect the display of the targeted symbol. However,
rather than updating symbol display (which would potentially have larger downstream impact), we'll
just include the option in this case.