Replace timestamp with hash in generated files
Some of our generators were embedding timestamps in the output files as comments.  This was indirectly preventing us from producing determistic builds as every rebuild came from different content.

Issue #7270 is tracking changing determinism to not be influenced by the comments.  Short term though changing this from using a time stamp (which changes and doesn't have meaning between machines) to outputing the checksum of the input file it ran against.