Move tests to inproc server
This moves the majority of the tests to an in-proc server.

Many are failing due to a race in the server.  Even though the
compilation task has completed and sent data to the command line the
task has not resolved in the server loop.  Cancellation of the server
ends up racing with that task completing.  Need to add a mode to the
server where it won't exit until the outstanding work is complete.