fixed flaky squiggle test
the reason it became flaky was that diagnostic analyzer events became async events to make sure events are serialized. and this test was written before the change and assumed events are synchronous.

due to layers of test code wrapped around another, this change didn't propagate enough up to find out this squiggle tests.

now we set async listener properly so it should become stable again.