Fixes that unverifiable IL is generated for switch over Nullable<Int64> and Nullable<UInt64>
The compiler tries to create a conversion of the expression on the case switch label to the type that is switched on. if the parameter is a Nullable this doesn't work and it assumed that the type is compatible with the underlying type of the Nullable expression. This is not the case if we have a Nullable<long> and an integer on the case switch.