Test shutdown in middle of compilation
Added a test to verify the shutdown operation is passive.  If any
existing compilations are in progress they will be allowed to complete
after the shutdown is received.

This required a bit of code refactoring.  It needed to be easier to
intercept compilation requests and block on them until a given set of
events had completed.